% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_wd.r
\name{append_wd}
\alias{append_wd}
\title{Append work directory paths}
\usage{
append_wd(wd = NULL, in_script = NULL, out_script = NULL)
}
\arguments{
\item{wd}{character Path to the work directory with the ALMaSS program}

\item{in_script}{character Path to the base script into which the WD is appended}

\item{out_script}{character File name for the modified base script. This will
usually be the same name as the base script.}
}
\description{
Append work directory paths to already excisting scripts
}
