% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AppendWorkDir.r
\name{AppendWorkDir}
\alias{AppendWorkDir}
\title{Append work directory paths}
\usage{
AppendWorkDir(WorkDir = NULL, InScript = NULL, OutScript = NULL)
}
\arguments{
\item{WorkDir}{character Path to the work directory with the ALMaSS exe file}

\item{InScript}{character Path to the base script into which the WD is appended}

\item{OutScript}{character File name for the modified base script. This will
usually be the same name as the base script.}
}
\description{
Append work directory paths to already excisting scripts
}

