% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_habuse_fit.r
\name{calc_habuse_fit}
\alias{calc_habuse_fit}
\title{Calculate the fit of habitat use frequencies}
\usage{
calc_habuse_fit(sim, obs)
}
\arguments{
\item{sim}{tibble The table with simulated data}

\item{obs}{tibble The table with the field data}
}
\value{
tibble The fit per species
}
\description{
Calculate the fit between habitat use scored in the field
and habitat use based on the simulations.
}
