% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcHabitatUseFit.r
\name{CalcHabitatUseFit}
\alias{CalcHabitatUseFit}
\title{Calculate the fit of habitat use frequencies}
\usage{
CalcHabitatUseFit(FieldData, SimData)
}
\arguments{
\item{FieldData}{data.table The table with the field data}

\item{SimData}{data.table The field forage output file from ALMaSS}
}
\value{
data.table The fit per species and month
}
\description{
Calculate the fit between habitat use scored in the field 
and habitat use based on the simulations
}

