% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcNoHunterFit.r
\name{CalcNoHunterFit}
\alias{CalcNoHunterFit}
\title{Calculate the fit to the number of hunters per farm}
\usage{
CalcNoHunterFit(respath = NULL, method = NULL)
}
\arguments{
\item{respath}{character Path to the result files}

\item{method}{character Method to calculate fit. Either 'overlap' or
'leastsq'.}
}
\value{
numeric The overlap
}
\description{
This function is only used to preform the test on scenarios already run.
The batchr.r scripts runs the test on the fly. Added to ralmass just to
keep track of things. Will probably be removed further down.
}

