% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPseudoCentroid.r
\name{CalcPseudoCentroid}
\alias{CalcPseudoCentroid}
\title{Calculate pseudo centroid for polygons}
\usage{
CalcPseudoCentroid(data, project = TRUE)
}
\arguments{
\item{data}{data.frame The raw file from the bird database}

\item{project}{logical Should the input coordinates be projected to UTM?}
}
\value{
data.frame A data.frame with the polygon id and the pseudo centroid
}
\description{
Calculate the mean x and mean y coordinate for a number of polygons
Be aware that this is only an approximation. With funny shaped polygons 
you might get strange results. Particularly if the vertices are not
are not evenly spaced.
}

