% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcWeightFit.r
\name{CalcWeightFit}
\alias{CalcWeightFit}
\title{Calculate the fit of the simulated weights to the weights observed}
\usage{
CalcWeightFit(SimData, FieldData)
}
\arguments{
\item{SimData}{data.table The file GooseEnergeticsData.txt}

\item{FieldData}{data.table The observed weights}

\item{Period}{character The period to calculate means within (currently only 
week is available )}
}
\description{
Calculate a least squares fit between field weights and simulated weights. 
The weights are scaled to go from 0 - 1 on the indivual observations.
A period of time must be chosen to calculate means within.
}

