% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDensity.r
\name{CheckDensity}
\alias{CheckDensity}
\title{Check densities of hunters per farm}
\usage{
CheckDensity(data = NULL, maxdensity = NULL, area = NULL)
}
\arguments{
\item{data}{data.table The file HuntingLocationsFarm.txt}

\item{maxdensity}{numeric The density paramter setting}

\item{area}{character The name of the column on which the area calculation
should be based.}
}
\value{
logical TRUE if all densities legal, FALSE otherwise
}
\description{
Do a check of the densities of hunters per farm. 
This is just ensuring that the density criteria is not broken.
}

