% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CheckFarmRefFreq.r
\name{CheckFarmRefFreq}
\alias{CheckFarmRefFreq}
\title{Check the frequency of farm reference numbers}
\usage{
CheckFarmRefFreq(resdir = NULL)
}
\arguments{
\item{resdir}{character The path to the hunting locations file}
}
\value{
data.table a table with the frequencies of individual farm
reference numbers.
}
\description{
Do a check of the frequencies of individual farm reference numbers.
The function summarise the frequency across the hunting locations and return
a table ready to plot.
}

