% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckRotation.r
\name{CheckRotation}
\alias{CheckRotation}
\title{Check crop rotations}
\usage{
CheckRotation(path = NULL)
}
\arguments{
\item{path}{character The path to the rotations to check. This can be the
run directory.}
}
\description{
Do a check of the crop rotations. There are certain crops that never should
be grown in to consecutive years for example. There is also a check two  
ensure that crops that are undersown other crops are valid combinations 
(some crops are never grown with an undersown crop)
The function will print the results of the check to the console.
}

