% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClassifyFields.r
\name{ClassifyFields}
\alias{ClassifyFields}
\title{Classify fields based on the last digit in the farm reference number}
\usage{
ClassifyFields(polyref, firstfield = 1e+05)
}
\arguments{
\item{polyref}{data.frame A dataframe which was produced using \code{\link{CleanAttrTable}}.}

\item{firstfield}{numeric The first elementtype number which is a field belonging to a farm.}
}
\value{
A data.frame with the same columns as in the input, but with the element types
corrected and the farmref stripped of its last digit (the field type identifier).
}
\description{
Classify fields based on the last digit in the farm reference number
}
\details{
List explaining the translation from one digit to the ALMaSS landscape element code:
\itemize{
  \item 0 = 20
  \item 1 = 27
  \item 2 = 33
  \item 3 = 35
  \item 4 = 55
  \item 5 = 56
  \item 6 = 59
  \item 7 = 60
  \item 8 = 71
  \item 9 = 68
 }
}

