% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClassifyHabitatUse.r
\name{ClassifyHabitatUse}
\alias{ClassifyHabitatUse}
\title{Classify animal habitat use}
\usage{
ClassifyHabitatUse(data, species = "Goose")
}
\arguments{
\item{data}{data.table A data.table with the habitat use from ALMaSS}

\item{species}{character String giving the species (or animal model)}
}
\value{
A data.table with the classifed habitat use.
}
\description{
Classify animal habitat use from what ALMaSS uses into categories
matching the ones that are scored in the field
}

