% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClassifyPreviousCrop.r
\name{ClassifyPrevCrop}
\alias{ClassifyPrevCrop}
\title{Classify Veg type based on the previous crop with respect to geese forage}
\usage{
ClassifyPrevCrop(PreviousCrop)
}
\arguments{
\item{PreviousCrop}{character A string with the name of the previous crop.}
}
\value{
The vegtype matching the one scored in the field
}
\description{
Classify the veg types from the detailed ones ALMaSS uses into categories
matching the ones that are scored in the field.
}
\details{
This function handles the special cases where crop type changes before the
actual farm management takes place in the simulation.
The function is designed to be used in combination with 'sapply'
}

