% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyVegType.r
\name{ClassifyVegType}
\alias{ClassifyVegType}
\title{Classify Veg type with respect to geese forage}
\usage{
ClassifyVegType(VegTypeCombo)
}
\arguments{
\item{VegTypeCombo}{character A string combined from VegType and VegPhase 
separated by '-'}
}
\value{
The vegtype matching the one scored in the field
}
\description{
Classify the veg types from the detailed ones ALMaSS uses into categories
matching the ones that are scored in the field
}
\details{
This function is designed to be used in combination with 'sapply'
}

