% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanAPIData.r
\name{CleanAPIData}
\alias{CleanAPIData}
\title{Clean the API field data}
\usage{
CleanAPIData(APIData)
}
\arguments{
\item{APIData}{data.table the raw field data}
}
\value{
The cleaned up data.
}
\description{
Clean API field data and calculate weights from API scores.
}

