% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertGameType.r
\name{ConvertGameType}
\alias{ConvertGameType}
\title{Convert numeric game type to character}
\usage{
ConvertGameType(GameType)
}
\arguments{
\item{GameType}{integer A game type}
}
\value{
The species name
}
\description{
Spell out the full name of the species shot from the numeric code
}

