% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertParam.r
\name{ConvertParam}
\alias{ConvertParam}
\title{Convert ALMaSS config parameter to a meaningful character string}
\usage{
ConvertParam(param = NULL)
}
\arguments{
\item{param}{character The input parameter}
}
\value{
character String with more meaningful explanation
}
\description{
Convert ALMaSS config parameter to a meaningful character string. This
can be useful when doing plots and generating labels and titles 
automatically.
}

