% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConvertSimDay.r
\name{ConvertSimDay}
\alias{ConvertSimDay}
\title{Supply the calendar date from a sim day}
\usage{
ConvertSimDay(DayNumber, Origin = "1990-01-01")
}
\arguments{
\item{DayNumber}{integer A day number}

\item{Origin}{Date Origin for Julian day calculation (YYYY-MM-DD)}
}
\value{
The calendar date
}
\description{
Return the calendar date from a simulation day number
}

