% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertVegPhase.r
\name{ConvertVegPhase}
\alias{ConvertVegPhase}
\title{Convert numeric vegphase to its character name}
\usage{
ConvertVegPhase(VegPhase)
}
\arguments{
\item{VegPhase}{integer A vegphase number}
}
\value{
The vegphase name
}
\description{
Spell out the full name of the vegphase from the numeric code
}

