% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_bat.r
\name{edit_bat}
\alias{edit_bat}
\title{Edit bat file or shell script}
\usage{
edit_bat(work_dir, bat_file)
}
\arguments{
\item{work_dir}{character Path to the work directory with the ALMaSS exe file}

\item{bat_file}{character File name for the shell file doing the loop}
}
\description{
Edit a shell script to reflect the correct number of runs
in a scenario. This function requires the ParameterValues.txt file
to be present in work_dir. The function reads the number of runs
from that file.
}
