% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditConfig.r
\name{EditConfig}
\alias{EditConfig}
\title{Edit config file}
\usage{
EditConfig(file = NULL, config = NULL, value = NULL)
}
\arguments{
\item{file}{character Full path to the work config file}

\item{config}{character The config to edit}

\item{value}{character The value of the config}
}
\description{
Edit a config file with a specific config variable and value. The value
simply added to the end of the file.
}

