% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditIni.r
\name{EditIni}
\alias{EditIni}
\title{Edit ini file}
\usage{
EditIni(WorkDir = NULL, Model = NULL, StandardProbes = TRUE,
  Batch = TRUE, nyear = 1)
}
\arguments{
\item{WorkDir}{character Path to the work directory with the TI_inifile}

\item{Model}{character The model to run}

\item{StandardProbes}{logical Run all the standard output probes?}

\item{Batch}{logical Edit BatchALMaSS.ini (cmdline) file (TRUE, default) or 
or the TI_inifile.ini (FALSE, GUI)}

\item{nyear}{numeric The number of years to run}
}
\description{
Edit the TI_inifile to reflect the species and output probes of choise
}

