% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EobsExtract.r
\name{ExtractEOBS}
\alias{ExtractEOBS}
\title{Extract values from E-OBS netCDF files}
\usage{
ExtractEOBS(eobs = NULL, spobject = NULL, metric = "mean")
}
\arguments{
\item{eobs}{character Vector of full paths to the eobs nc files}

\item{spobject}{Spatial* The geometry to extract with (point of polygon)}

\item{metric}{character The metric to use when spoject is a polygon.  Will
normally be 'mean', 'min' or 'max'}
}
\description{
Extract values based on a spatial object form the E-OBS data.
Useful when making weather files for ALMaSS. Often a point in the center
of the landscape can be used to extract values. For the bigger landscapes
a mean for the landscapes might make more sense. The function can handle
that too.
}

