% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest.r
\name{find_closest}
\alias{find_closest}
\title{find_closest finds the nearest roost}
\usage{
find_closest(x, y, species)
}
\arguments{
\item{x}{int The x-coordinate of the forage location}

\item{y}{int The y-coordinate of the forage location}

\item{species}{character The species name}
}
\value{
double The distance to the nearest roost in meters
}
\description{
The find_closest function returns the distance to the nearst roost for
each species. Designed to be used with purr::pmap
}
