% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FixSeparator.r
\name{FixSeparator}
\alias{FixSeparator}
\title{Remove comma as thousand separator}
\usage{
FixSeparator(dt = NULL, column = NULL)
}
\arguments{
\item{dt}{data.table The data.table with the column containing commas.}

\item{column}{character Name of the column with commas.}
}
\value{
The data.table with the comma issue fixed.
}
\description{
Some tables get exported from ArcGis using commas as thousand separator
and are therefore imported as character. This function simply removes the
commas.
}

