% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenerateParams.r
\name{GenerateParams}
\alias{GenerateParams}
\title{Generate a range of paramter values}
\usage{
GenerateParams(start, end, N, Config)
}
\arguments{
\item{start}{numeric Starting value.}

\item{end}{numeric End value.}

\item{N}{Number of required values.}
}
\value{
data.frame Two column data.frame. First column the config variable
and second column being the values.
}
\description{
Generate a range of paramter values from specified start
and end values. A simple wrapper around seq
}

