% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_func_response.r
\name{get_func_resp}
\alias{get_func_resp}
\title{get_func_resp
Return value from functional reponse}
\usage{
get_func_resp(dens = NULL, resp = NULL, species = NULL)
}
\arguments{
\item{dens}{double The grain density}

\item{resp}{data.table The curve points written from ALMaSS}

\item{species}{character Either Pinkfoot, Barnacle or Greylag}
}
\value{
double The response.
}
\description{
The geese have different functional responses to grain density.
This function return the repsonse given a grain density.
}
