% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_numbers.R
\name{get_input_numbers}
\alias{get_input_numbers}
\title{Get total number of geese entering the simulation}
\usage{
get_input_numbers(config)
}
\arguments{
\item{config}{character The config variable as read in by readLines}
}
\value{
tibble A table with the calculated numbers
}
\description{
The actual number of geese in the simulation is a function of multiple
parameters. This function reads these from the config file and computes
the numbers.
}
