% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_month.R
\name{get_month}
\alias{get_month}
\title{Get month from julian date or day in year}
\usage{
get_month(day)
}
\arguments{
\item{day}{integer The simulation day or day in year from ALMaSS}
}
\value{
month as integer
}
\description{
Many ALMaSS outputs has the simulation day number as it's only time indicator.
Often one would want to compute or plot stuff using dates. This function
translates from either julian date or day in year to month.
}
