% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParamString.r
\name{GetParamString}
\alias{GetParamString}
\title{Get the string of a config variable}
\usage{
GetParamString(config = NULL)
}
\arguments{
\item{config}{character String from which to extract the config parameter
string.}
}
\value{
The character string of the config parameter
}
\description{
Return the character string of a config variable in a specified
 ALMaSS config file.
}

