% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParamType.r
\name{GetParamType}
\alias{GetParamType}
\title{Add type to input parameter string}
\usage{
GetParamType(param = NULL)
}
\arguments{
\item{param}{character The input paramter}
}
\value{
character param plus the type added in parantheses
}
\description{
Add the input parameter type to the string of just the parameter name.
}

