% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParamValue.r
\name{GetParamValue}
\alias{GetParamValue}
\title{Get the value of a config variable}
\usage{
GetParamValue(config = NULL, param = NULL)
}
\arguments{
\item{config}{character Path to the file to read start and leave date from}

\item{param}{character The name of the config variables}
}
\value{
The value of the config parameter
}
\description{
Return the value of a config variable in a specified ALMaSS config file.
}

