% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotAOR.r
\name{PlotAOR}
\alias{PlotAOR}
\title{Plot the AOR index}
\usage{
PlotAOR(data, x = NULL, y = NULL, scenarios = NULL, fixed = TRUE,
  title = NULL, triangle = FALSE, trigger = 0)
}
\arguments{
\item{data}{data.frame Data.frame with x, y and scenario column.}

\item{x}{character The name of the column with x coordinates (occupancy).}

\item{y}{character The name of the column with y coordinates (abundance).}

\item{scenarios}{character The name of the column indicating which scenario
a given set of coordinates belong to.}

\item{fixed}{logical If TRUE (the default) plot has xlim and ylim -1,1.}

\item{title}{character The title of the plot (optional).}

\item{triangle}{logical Should the area of decrease on both axes be highlighted?
default is FALSE.}

\item{trigger}{numeric The trigger value for accepted decrease in the
summed-AOR indexs score. Mostly used when assessing impact of pesticides.
Any points falling beyond the area highlighted by the shaded triangle are an
unacceptable level of impact.}
}
\value{
An AOR plot
}
\description{
Plot the AOR index (Hoye et al. 2012). The index is based on the
abundance-occupancy relationship and facilitates the interpretation of
outputs from individual based models.
}
\examples{
df = data.frame('x' = rnorm(8, 0, .5), 'y' = rnorm(8, 0, .5), 'scenario' = letters[1:8])
PlotAOR(df, x = 'x', y = 'y', scenarios = 'scenario')
}
\references{
Hoye, T. T., Skov, F. & Topping, C.J. (2012). Interpreting
outputs of agent-based models using abundance-occupancy relationships.
Ecological Indicators 20: 221-227.
}

