% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotGooseForage.r
\name{PlotGooseForage}
\alias{PlotGooseForage}
\title{Plot forage available to geese}
\usage{
PlotGooseForage(data, species = "Pinkfoot", type = NULL,
  plottype = "points", dates = TRUE)
}
\arguments{
\item{data}{data.frame or data.table The ALMaSS data file GooseFieldForageData.txt}

\item{species}{character One of the goose species ('Pinkfoot', 'Barnacle', 'Greylag') defaults to 'Pinkfoot'}

\item{type}{character What forage should be plotted ('Grass', 'Grain' or 'Combined'). No default.}
}
\value{
A nice plot
}
\description{
Plot the forage available as grazing and grain available to geese in ALMaSS
}

