% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotProbe0}
\alias{PlotProbe0}
\title{Plot results from probe0}
\usage{
PlotProbe0(data, seasons = TRUE, lty = "l", add = FALSE,
  species = "Dormouse")
}
\arguments{
\item{data}{data.frame The raw output from probe0 or probe from ALMaSS}

\item{seasons}{logical Should breeding season and hibernation polygons be added to the plot?
(dormouse specific)}

\item{lty}{character Either 'l' = lines, 'b' both lines and point.}

\item{add}{Should the plot be added to an already existing plot in the plot window?}

\item{species}{What species is the plot for? (currenly only "Dormouse" and "Goose" available)}
}
\value{
A nice plot
}
\description{
Plot the results from probe0
Note: you should read the results file using the fread function in the data.table package
It is very fast - you will not regret it.
}

