% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTerritoryProbe.r
\name{PlotTerritoryProbe}
\alias{PlotTerritoryProbe}
\title{Plot results from the TerritoryProbe}
\usage{
PlotTerritoryProbe(data, seasons = TRUE, type = "l")
}
\arguments{
\item{data}{data.frame The raw output from the TerritoryProbe or probe from ALMaSS}

\item{seasons}{logical Should breeding season and hibernation polygons be added to the plot?}

\item{type}{character Either 'l' = lines, 'b' = both lines and point.}
}
\value{
A nice plot
}
\description{
Plot the results from the TerritoryProbe (currently only implemented in the Dormouse code)
}

