% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotVegDump.r
\name{PlotVegDump}
\alias{PlotVegDump}
\title{Plot ouput from VegDump}
\usage{
PlotVegDump(data)
}
\arguments{
\item{data}{data.table The raw VegDump.txt from ALMaSS}
}
\value{
A nice plot
}
\description{
Plot ouput from VegDump.txt
}
\details{
The VegDump probe output various parameters from the landscape
Look in the actual file to see what is currently implemented

Note: you should read the results file using the fread function in the data.table package
It is very fast - you will not regret it.
}

