% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeather.R
\name{PlotWeather}
\alias{PlotWeather}
\title{Plot an Almass weather file}
\usage{
PlotWeather(file, from = NULL, to = NULL)
}
\arguments{
\item{file}{character The full path to the weather file}

\item{from}{character Date in format year-month-day}

\item{to}{character Date in format year-month-day}
}
\value{
A nice plot
}
\description{
Plot an Almass weather file or a subset of it
}
\details{
If from or to are left blank the entire range of data is plotted
}

