% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RuleSetPrint.r
\name{RuleSetPrint}
\alias{RuleSetPrint}
\title{Print rule set number and description}
\usage{
RuleSetPrint(ruleset = NULL)
}
\arguments{
\item{ruleset}{numeric The rule set number}
}
\value{
character The rule set number and description
}
\description{
Print rule set number and description
}

