% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoreResults.r
\name{StoreResults}
\alias{StoreResults}
\title{Copy and datestamp result files}
\usage{
StoreResults(InDirectory = NULL, OutDirectory = NULL)
}
\arguments{
\item{InDirectory}{character Path to the run directories}

\item{OutDirectory}{character Path to the location where the files will be
stored.}
}
\description{
This function copies and renames the ParameterFittingResults.txt
file to a user specified location. The original file name will padded by
the parent directory name (leading) and the output from Sys.Date() (trailing)
}

