% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SummarizeCIPE.r
\name{SummarizeCIPE}
\alias{SummarizeCIPE}
\title{Summarize output from the CIPEgrid probe}
\usage{
SummarizeCIPE(CIPEgrid, plot = TRUE)
}
\arguments{
\item{CIPEgrid}{data.frame A dataframe in the raw format from the CIPEgrid probe.}
}
\value{
A data.frame with two columns. One is the timestep, the other the density
}
\description{
Calculate the average density of animals in the landscape based on output
from the CIPEgrid probe.
}

