% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummaryAttrTable.r
\name{SummaryAttrTable}
\alias{SummaryAttrTable}
\title{Summary statistics for an attribute table}
\usage{
SummaryAttrTable(AttrTable)
}
\arguments{
\item{AttrTable}{data.frame An attribute table cleaned using \code{\link{CleanAttrTable}}}
}
\value{
Prints the results to the screen
}
\description{
Calculate some simple statistics based on the attribute table.
The purpose of this function is mostly to enable easy checks of things
like total cell numbers and types of elements.
}

