% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_aor.r
\name{tidy_aor}
\alias{tidy_aor}
\title{Tidy the standard AOR output}
\usage{
tidy_aor(data)
}
\arguments{
\item{data}{tibble The AOR table from ALMaSS}
}
\value{
a tibble
}
\description{
The standard AOR output has a format thats not easy
to program with. This function transforms it into
a tidy format.
}
