% Generated by roxygen2 (4.0.2): do not edit by hand
\name{UtmToALMaSS}
\alias{UtmToALMaSS}
\title{Convert UTM coordinates to ALMaSS map coordinates}
\usage{
UtmToALMaSS(data, long, lat, map)
}
\arguments{
\item{data}{data.frame The data frame with the coordinates}

\item{long}{character The name of the column holding the longitude}

\item{lat}{character The name of the column holding the latitude}

\item{map}{character The name of the map used in ALMaSS}
}
\value{
data.frame A data.frame with the ALMaSS coordinates
}
\description{
Convert UTM coordinates into coordinates that can be used in
in ALMaSS directly. Useful when importing field data into ALMaSS
}
\details{
For Vejlerne there are currently two working maps: VejlerneTest
(10000*7000) and VejlerneBigMap (36000*24000).
}

