% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteAlmassInput.r
\name{WriteAlmassInput}
\alias{WriteAlmassInput}
\title{Write input file to ALMaSS}
\usage{
WriteAlmassInput(table = NULL, pathtofile = NULL, headers = TRUE)
}
\arguments{
\item{table}{data.table The table with the input}

\item{pathtofile}{character The path to the where the file is 
written (including the name and extension of the file, usually .txt)}

\item{headers}{logical Should the output file include headers?}
}
\value{
A tab separated text file formatted as an ALMaSS input file.
}
\description{
Write an input file ready for loading in ALMaSS. This has the format of line
1 being the number of rows, line 2 the headers (if needed) and then the 
actual content of the colummns. The file is a tab delimeted txt file.
}
\details{
A few of the older ALMaSS input files do not contain headers. This can be 
switched on or of using the headers argument.
}
