% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteReclassTable.r
\name{WriteReclassTable}
\alias{WriteReclassTable}
\title{Write reclass table}
\usage{
WriteReclassTable(Table, PathToFile)
}
\arguments{
\item{Table}{data.table A two column table where the first column is the 
polyref number from the ALMaSS landscape. The second column is the entity
to be mapped e.g goose numbers.}

\item{PathToFile}{character The path to the where the file is 
written (including the name and extension of the file, .txt)}
}
\description{
Write a reclassification table to be used with the arcpy tool
 ReclassByASCIIFile.
}

