% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDistToRoosts.r
\name{CalcDistToRoosts}
\alias{CalcDistToRoosts}
\title{Calculate the distance from fields to goose roosts}
\usage{
CalcDistToRoosts(roost = NULL, fields = NULL, fieldobs = NULL,
  polyref = NULL, species = NULL)
}
\arguments{
\item{roost}{data.table The GooseRoosts.txt file}

\item{fields}{data.table If fieldobs = FALSE then the GooseFieldForageData.txt
 file otherwise a data.table with field observations converted to ALMaSS 
coordinates using the UtmToAlmass function.}

\item{fieldobs}{logical If TRUE then input is fieldobservations otherwise
output from ALMaSS simulations}

\item{polyref}{data.table The object with the polyref file (only if
fieldobs = FALSE)}

\item{species}{character A vector of species names for which the distances should be 
calculated for. Either 'Barnacle', 'Pinkfoot' or 'Greylag'}
}
\value{
data.table A data.table with the polyrefnumber or if fieldobs then
 PolyID, distances for each roost and a column with the shortest distance.
}
\description{
Calculate the distance from fields to goose roosts (in this case)
Using the centroids from the polyref file distances can be calculated 
outside of ALMaSS.
}

