% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalcDistToRoosts.r
\name{CalcDistToRoosts}
\alias{CalcDistToRoosts}
\title{Calculate the distance from fields to goose roosts}
\usage{
CalcDistToRoosts(roost, fields, polyref, species)
}
\arguments{
\item{roost}{data.table The object with the GooseRoosts.txt file}

\item{fields}{data.table The object with the GooseFieldForageData.txt file}

\item{polyref}{data.table The object with the polyref file}

\item{species}{character The species for which the distances should be
calculated for. Either 'Barnacle', 'Pinkfoot' or 'Greylag'}
}
\value{
list A list holding a data.table with distances for each roost
}
\description{
Calculate the distance from fields to goose roosts (in this case)
Using the centroids from the polyref file distances can be calculated
outside of ALMaSS.
}

