% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcFlockSizeFit.r
\name{CalcFlockSizeFit}
\alias{CalcFlockSizeFit}
\title{Calculate the fit of the simulated flock size distribution to the observed
flock size distribution}
\usage{
CalcFlockSizeFit(sim = NULL, obs = NULL)
}
\arguments{
\item{sim}{numeric A vector of flock sizes from the simulation}

\item{obs}{tibble  A tibble with a columns identifying the species
and a column with each flock observation from the simulation.}
}
\value{
numeric The calculated fit.
}
\description{
Calculate fit between the simulated flock size distribution to the observed
flock size distribution based as 1 minus the least squares difference.
}
\details{
The fit is measured on the proportions of the flock sizes falling into
each bin of 10% quantiles of the observed range of flock sizes.
The fit is in both cases measured over the entire season.
}
