% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CalcPseudoCentroid}
\alias{CalcPseudoCentroid}
\title{Calculate pseudo centroid for polygons}
\usage{
CalcPseudoCentroid(data)
}
\arguments{
\item{data}{data.frame The raw file from the bird database}
}
\value{
data.frame A data.frame with the polygon id and the pseudo centroid
}
\description{
Calculate the mean x and mean y coordinate for a number of polygons
Be aware that this is only an approximation. With funny shaped polygons
you might get strange results. Particularly if the vertices are not
are not evenly spaced.
}

