% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcWeightFit.r
\name{CalcWeightFit}
\alias{CalcWeightFit}
\title{Calculate the fit of the simulated weights to the weights observed}
\usage{
CalcWeightFit(Sim = NULL, Field = NULL, measure = NULL)
}
\arguments{
\item{Sim}{data.table The file GooseEnergeticsData.txt}

\item{Field}{data.table The observed weights}

\item{measure}{character Fit measure to use. Either LS or SSSE}
}
\description{
Calculate fit between field weights and simulated weights using either least 
squares or the sum of standardized squared errors (SSSE, Frank & Baret 2013)
}
\references{
Frank, B. M. and P. V. Baret (2013). "Simulating brown trout
 demogenetics in a river/nursery brook system: The individual-based
 model DemGenTrout." Ecological Modelling 248: 184-202.
}

