% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CheckDensity.r
\name{CheckDensity}
\alias{CheckDensity}
\title{Check densities of hunters per farm}
\usage{
CheckDensity(data = NULL, MaxDensity = NULL)
}
\arguments{
\item{data}{data.table The file HuntingLocationsFarm.txt}

\item{MaxDensity}{numeric The density paramter setting}
}
\value{
logical TRUE if all densities legal, FALSE otherwise
}
\description{
Do a check of the densities of hunters per farm.
This is just ensuring that the density criteria is not broken.
}

