% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditBat.r
\name{EditBat}
\alias{EditBat}
\title{Edit bat file or shell script}
\usage{
EditBat(WorkDir = NULL)
}
\arguments{
\item{WorkDir}{character Path to the work directory with the ALMaSS exe file}
}
\description{
Edit a bat file or shell script to reflect the correct number of runs
in a scenario. This function requires the ParameterValues.txt file
to be present in WorkDir. The function reads the number of runs
from that file.
}
