% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditHunterInput.r
\name{EditHunterInput}
\alias{EditHunterInput}
\title{Edit input file for the Hunter model}
\usage{
EditHunterInput(hhlpath = NULL, column = NULL, change = NULL,
  weekbehav = 0)
}
\arguments{
\item{hhlpath}{character Path to an existing Hunter_Hunting_Locations.txt file}

\item{column}{character Name of column to apply change to}

\item{change}{numeric The change to apply. See details}

\item{weekbehav}{numeric The weekly hunting behaviour. See details}
}
\value{
A tab separated text file formatted as an ALMaSS input file.
}
\description{
Some of the input parameters for the hunter model are given as individual
values to each object. This is controlled via the
 Hunter_Hunting_Locations.txt file and this \code{EditHunterInput} will
apply changes to specified inputs.
}
\details{
\code{change} gives the proportion of hunters which will have a 1 in 
WeekdayHunterChance or GooseLookChance. When \code{column} is HuntingDays
\code{change} will increase (when \code{change} > 1) or decrease 
(when \code{change} < 1) the already existing distribution of HuntingDays. 
When \code{column} is Efficiency, \code{change} will be the actual value
given to all hunters.

\code{weekbehav} controls the weekly hunting behaviour. 0 is weekend hunting
only (default), 1 is not weekend only and 2 means leave a refraction period
 in between hunting events. The refraction period is controlled by a config
 variable in ALMaSS.
}

