% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditHunterInput.r
\name{EditHunterInput}
\alias{EditHunterInput}
\title{Write input file to ALMaSS}
\usage{
EditHunterInput(hhl = NULL, column = NULL, change = NULL)
}
\arguments{
\item{hhl}{data.table An existing Hunter_Hunting_Locations.txt file}

\item{column}{character Name of column to apply change to}

\item{change}{numeric The change to apply. See details}
}
\value{
A tab separated text file formatted as an ALMaSS input file.
}
\description{
Write an input file ready for loading in ALMaSS. This has the format of line
1 being the number of rows, line 2 the headers (if needed) and then the 
actual content of the colummns. The file is a tab delimeted txt file.
A few of the older ALMaSS input files do not contain headers. This can be 
switched on or of using the headers argument.
}
\details{
\code{change} gives the proportion of hunters which will have a 1 in 
WeekdayHunterChance or GooseLookChance. When \code{column} is HuntingDays
\code{change} will increase (when \code{change} > 1) or decrease 
(when \code{change} < 1) the already existing distribution of HuntingDays. 
When \code{column} is Efficiency, \code{change} will be the actual value
given to all hunters.
}

