% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenerateParams.r
\name{GenerateParams}
\alias{GenerateParams}
\title{Generate a range of paramter values}
\usage{
GenerateParams(start, end, N, Config, start2 = NULL, end2 = NULL,
  N2 = NULL, Config2 = NULL, write = FALSE)
}
\arguments{
\item{start}{numeric Starting value.}

\item{end}{numeric End value.}

\item{N}{Number of required values.}

\item{Config}{The character string with the config variable (including (int))}

\item{start2}{numeric Starting value for the (optional) second variable}

\item{end2}{numeric End value for the (optional) second variable}

\item{N2}{Number of required values for the (optional) second variable}

\item{Config2}{The character string with the second config variable (including (int))}
}
\value{
data.frame One column each value a text string with the config and
and its value
}
\description{
Generate a range of paramter values from specified start
and end values. A simple wrapper around seq
}

