% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateParams.r
\name{GenerateParams}
\alias{GenerateParams}
\title{Generate a range combinations of paramter values}
\usage{
GenerateParams(..., write = FALSE, splits = NULL)
}
\arguments{
\item{...}{numeric One or more vectors (usually made with seq and the
length.out argument)}

\item{write}{logical Should a file with the table be written to disk?}

\item{splits}{numeric The number of files to split the scenarios into. 
(Optional). If not NULL the files gets written to disk.}
}
\value{
data.frame One column data.frame where each line is a text string
with the config variable and its value
}
\description{
Generate all possible combinations of values in the input vectors and 
(optionally) write them to a file. Typically used when making scenarios for 
fitting parameters.
}
\details{
This function is a modification of expand.grid from base.
}
\examples{
val = seq(1, 5, length.out = 5)
val2 = seq(10, 50, length.out = 5)
GenerateParams('GOOSE_MINFORAGEOPENNESS' = val,
'HUNTERS_MAXDENSITY' = val2, write = FALSE)
}

