% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEnergetics.r
\name{PlotEnergetics}
\alias{PlotEnergetics}
\title{Plot energy output from the ALMaSS Goose Management model}
\usage{
PlotEnergetics(SimData, FieldData, Sample = NULL)
}
\arguments{
\item{SimData}{data.table The raw output from the file GooseEnergeticsData.txt}

\item{FieldData}{data.table object cleaned by the function CleanAPIData}

\item{Sample}{numeric The proportion of the simulated data to include. Value
between 0.01 and 1.}
}
\value{
A nice plot
}
\description{
Plot the results from the energy probe
}
\details{
The energetics output can be very large, so please use the fread function 
 when loading the data. Currently only implemented for pinkfeet due to lack
 of field data on barnacle and greylag.
}

