% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotEnergetics.r
\name{PlotEnergetics}
\alias{PlotEnergetics}
\title{Plot energy output from the ALMaSS Goose Management model}
\usage{
PlotEnergetics(data, species = "all", package = "ggplot2", scales = NULL)
}
\arguments{
\item{data}{data.table The raw output from the file GooseEnergeticsData.txt}

\item{species}{character What species to plot? Defaults to "all"}

\item{package}{character Base- or ggplot2-type ? Defaults to "ggplot2"}

\item{type}{character Currently only accepts "mean".}
}
\value{
A nice plot
}
\description{
Plot the results from the energy probe
}
\details{
The energetics output can be very large, so please use the fread function
 when loading the data.
}

