% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotGooseForage.r
\name{PlotGooseForage}
\alias{PlotGooseForage}
\title{Plot forage available to geese}
\usage{
PlotGooseForage(data)
}
\arguments{
\item{data}{data.frame or data.table The ALMaSS data file GooseFieldForageData.txt}
}
\value{
A nice plot
}
\description{
Plot the forage available as grazing and grain available to geese in ALMaSS
}

