% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGooseHabitatUse.r
\name{PlotGooseHabitatUse}
\alias{PlotGooseHabitatUse}
\title{Plot goose habitat use}
\usage{
PlotGooseHabitatUse(data, species = NULL, variable = NULL)
}
\arguments{
\item{data}{data.frame or data.table The ALMaSS data file GooseFieldForage.txt}

\item{species}{character The species to plot. Either 'Barnacle', 'Pinkfoot' or 'Greylag'}

\item{variable}{character What variable are we plotting? Either 'VegTypeChr'
or 'HabitatUse'.}
}
\value{
A nice plot
}
\description{
Plot the frequency of use of particular habitat types
}

