% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGoosePopulation.r
\name{PlotGoosePopulation}
\alias{PlotGoosePopulation}
\title{Plot goose population data}
\usage{
PlotGoosePopulation(data, dates = FALSE)
}
\arguments{
\item{data}{data.table The ALMaSS data file GoosePopulationData.txt}

\item{dates}{logical Should we plot day numbers as dates?}
}
\value{
A nice plot
}
\description{
Plot the number of geese per species and type
}

