% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotGoosePopulation.r
\name{PlotGoosePopulation}
\alias{PlotGoosePopulation}
\title{Plot goose population data}
\usage{
PlotGoosePopulation(data)
}
\arguments{
\item{data}{data.frame or data.table The ALMaSS data file GoosePopulationData.txt}
}
\value{
A nice plot
}
\description{
Plot the number of geese per species and type
}

