% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PlotProbe0}
\alias{PlotProbe0}
\title{Plot results from probe0}
\usage{
PlotProbe0(data, seasons = TRUE, lty = "l")
}
\arguments{
\item{data}{data.frame The raw output from probe0 or probe from ALMaSS}

\item{seasons}{logical Should breeding season and hibernation polygons be added to the plot?}

\item{lty}{character Either 'l' = lines, 'b' both lines and point.}
}
\value{
A nice plot
}
\description{
Plot the results from probe0
}

