% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeather.R
\name{PlotWeather}
\alias{PlotWeather}
\title{Plot an Almass weather file}
\usage{
PlotWeather(file, subset = NULL)
}
\arguments{
\item{file}{character The full path to the weather file}

\item{subset}{numeric Either a year or a sequence of years. If left blank - plot the whole thing.}
}
\value{
A nice plot
}
\description{
Plot an Almass weather file or a subset of it
}

