% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/polyref.r
\docType{data}
\name{polyref}
\alias{polyref}
\title{Example of an attribute table exported from ArcGIS}
\format{A data frame with 1000 rows and 4 variables}
\usage{
polyref
\itemize{
  \item OBJECTID. This is the polygon ID.
  \item VALUE. So is this. A duplicate of OBJECTID.
  \item COUNT. The number of cells within the polygon.
  \item LINK. The type of landscape element.
 }
}
\description{
The table contains 1000 rows and 4 columns
}

