% Generated by roxygen2 (4.0.2): do not edit by hand
\name{UtmToALMaSS}
\alias{UtmToALMaSS}
\title{Convert UTM coordinates to ALMaSS map coordinates}
\usage{
UtmToALMaSS(data, long, lat, map, subset = TRUE, toalmass = TRUE)
}
\arguments{
\item{data}{data.frame The data frame with the coordinates}

\item{long}{character The name of the column holding the longitude}

\item{lat}{character The name of the column holding the latitude}

\item{map}{character The name of the map used in ALMaSS}

\item{subset}{logical Should the output only contain locations within the
ALMaSS map? Ignored when toalmass = FALSE}

\item{toalmass}{logical Do you want to convert to ALMaSS coordinates. Use FALSE if
if you want to convert ALMaSS map coordinates back to UTM
specified map?}
}
\value{
data.frame A data.frame with the ALMaSS coordinates
}
\description{
Convert UTM coordinates into coordinates that can be used in
in ALMaSS directly. Useful when importing field data into ALMaSS
}
\details{
For Vejlerne there are currently two working maps: VejlerneTest
(10000*7000) and VejlerneBigMap (36000*24000).
}

