% Generated by roxygen2 (4.0.2): do not edit by hand
\name{WritePolyref}
\alias{WritePolyref}
\title{Write polyref file}
\usage{
WritePolyref(Table, PathToFile)
}
\arguments{
\item{Table}{data.frame or data.table An attribute table}

\item{PathToFile}{The path to the where the file is written (including the name and extension of the file, usually .txt)}
}
\value{
A tab separated text file formatted according to the requirements for a polygon reference file for ALMaSS.
}
\description{
Write a polyref file to a text file ready for loading in ALMaSS.
}

